chore: cleanup graph colors and move to shared location
