+++ src/app/calculator/steam/saturated-properties/saturated-properties-chart/saturated-properties-chart.component.ts
@@ -2,7 +2,7 @@ import { Component, Input, ViewChild, ElementRef, HostListener, SimpleChanges, O
 import { Settings } from '../../../../shared/models/settings';
 import { SaturatedPropertiesOutput } from '../../../../shared/models/steam/steam-outputs';
 import { SimpleChart } from '../../../../shared/models/plotting';
-import { graphColors } from '../../../../phast/phast-report/report-graphs/graphColors';
+import { graphColors } from '../../../../shared/graphColors';
 
 import { SaturatedPropertiesService, IsobarCoordinates } from '../../saturated-properties.service';
 import { SaturatedPropertiesConversionService } from '../../saturated-properties-conversion.service';
