+++ src/app/calculator/steam/saturated-properties/saturated-properties-ph-chart/saturated-properties-ph-chart.component.ts
@@ -3,7 +3,7 @@ import { Settings } from '../../../../shared/models/settings';
 import { SimpleChart } from '../../../../shared/models/plotting';
 import { SaturatedPropertiesOutput } from '../../../../shared/models/steam/steam-outputs';
 import { SaturatedPropertiesConversionService } from '../../saturated-properties-conversion.service';
-import { graphColors } from '../../../../phast/phast-report/report-graphs/graphColors';
+import { graphColors } from '../../../../shared/graphColors';
 
 import { SaturatedPropertiesService, IsothermCoordinates } from '../../saturated-properties.service';
 import { PlotlyService } from 'angular-plotly.js';
