--- src/app/phast/phast-report/report-graphs/graphColors.ts
@@ -1,57 +1,0 @@
-// export const graphColors: Array<any> = [
-//     '#BA4A00',
-//     '#DC7633',
-//     '#BB8FCE',
-//     '#E74C3C',
-//     '#F9E79F',
-//     '#212F3C',
-//     '#4A235A',
-//     '#CA6F1E',
-//     '#F39C12',
-//     '#F1C40F',
-//     '#7B241C',
-//     '#909497',
-//     '#D2B4DE',
-//     "#512E5F",
-//     "#17A589",
-//     "#F4F6F7",
-//     "#117A65",
-//     "#9B59B6",
-//     "#17202A",
-//     "#641E16",
-//     "#D4EFDF"
-// ]
-export const graphColors: Array<any> = [
-  '#1E7640',
-  '#2ABDDA',
-  '#84B641',
-  '#7030A0',
-  '#E1CD00',
-  '#306DBE',
-  '#A03123',
-  '#7FD7E9',
-  '#DE762D',
-  '#948A54',
-  '#A9D58B',
-  '#FFE166',
-  '#DD7164',
-  '#3f4a7d'
-];
-
-
-export const phastGraphColors: Array<any> = [
-  '#1E7640',
-  '#2ABDDA',
-  '#84B641',
-  '#7030A0',
-  '#E1CD00',
-  '#306DBE',
-  '#A03123',
-  '#948A54',
-  '#DE762D',
-  '#1E7640',
-  '#1E7640',
-  '#FFE166',
-  '#DD7164',
-  '#3f4a7d'
-];
