+++ src/app/report-rollup/phast-rollup/phast-rollup.component.ts
@@ -3,7 +3,7 @@ import { Settings } from '../../shared/models/settings';
 import { Calculator } from '../../shared/models/calculators';
 import { PhastResultsData } from '../report-rollup-models';
 import { PieChartDataItem } from '../rollup-summary-pie-chart/rollup-summary-pie-chart.component';
-import { graphColors } from '../../phast/phast-report/report-graphs/graphColors';
+import { graphColors } from '../../shared/graphColors';
 import { ConvertUnitsService } from '../../shared/convert-units/convert-units.service';
 import * as _ from 'lodash';
 import { BarChartDataItem } from '../rollup-summary-bar-chart/rollup-summary-bar-chart.component';
