+++ src/app/shared/graphColors.ts
@@ -1,0 +1,150 @@
+export const graphColors: Array<any> = [
+  '#1E7640',
+  '#2ABDDA',
+  '#84B641',
+  '#7030A0',
+  '#E1CD00',
+  '#306DBE',
+  '#A03123',
+  '#7FD7E9',
+  '#DE762D',
+  '#948A54',
+  '#A9D58B',
+  '#FFE166',
+  '#DD7164',
+  '#3f4a7d'
+];
+
+
+//flat colors
+export const FlatColors = [
+    '#F2D7D5',
+    '#E6B0AA',
+    '#D98880',
+    '#CD6155',
+    '#C0392B',
+    '#A93226',
+    '#922B21',
+    '#7B241C',
+    '#641E16',
+    '#F5B7B1',
+    '#F1948A',
+    '#EC7063',
+    '#E74C3C',
+    '#CB4335',
+    '#B03A2E',
+    '#943126',
+    '#78281F',
+    '#D7BDE2',
+    '#C39BD3',
+    '#AF7AC5',
+    '#9B59B6',
+    '#884EA0',
+    '#76448A',
+    '#633974',
+    '#512E5F',
+    '#D2B4DE',
+    '#BB8FCE',
+    '#A569BD',
+    '#8E44AD',
+    '#7D3C98',
+    '#6C3483',
+    '#5B2C6F',
+    '#4A235A',
+    '#A9CCE3',
+    '#7FB3D5',
+    '#5499C7',
+    '#2980B9',
+    '#2471A3',
+    '#1F618D',
+    '#1A5276',
+    '#154360',
+    '#AED6F1',
+    '#85C1E9',
+    '#5DADE2',
+    '#3498DB',
+    '#2E86C1',
+    '#2874A6',
+    '#21618C',
+    '#1B4F72',
+    '#A3E4D7',
+    '#76D7C4',
+    '#48C9B0',
+    '#1ABC9C',
+    '#17A589',
+    '#148F77',
+    '#117864',
+    '#0E6251',
+    '#A2D9CE',
+    '#73C6B6',
+    '#45B39D',
+    '#16A085',
+    '#138D75',
+    '#117A65',
+    '#0E6655',
+    '#0B5345',
+    '#A9DFBF',
+    '#7DCEA0',
+    '#52BE80',
+    '#27AE60',
+    '#229954',
+    '#1E8449',
+    '#196F3D',
+    '#145A32',
+    '#ABEBC6',
+    '#82E0AA',
+    '#58D68D',
+    '#2ECC71',
+    '#28B463',
+    '#239B56',
+    '#1D8348',
+    '#186A3B',
+    '#F9E79F',
+    '#F7DC6F',
+    '#F4D03F',
+    '#F1C40F',
+    '#D4AC0D',
+    '#B7950B',
+    '#9A7D0A',
+    '#7D6608',
+    '#FAD7A0',
+    '#F8C471',
+    '#F5B041',
+    '#F39C12',
+    '#D68910',
+    '#B9770E',
+    '#9C640C',
+    '#7E5109',
+    '#F5CBA7',
+    '#F0B27A',
+    '#EB984E',
+    '#E67E22',
+    '#CA6F1E',
+    '#AF601A',
+    '#935116',
+    '#784212',
+    '#EDBB99',
+    '#E59866',
+    '#DC7633',
+    '#D35400',
+    '#BA4A00',
+    '#A04000',
+    '#873600',
+    '#6E2C00',
+    '#AEB6BF',
+    '#85929E',
+    '#5D6D7E',
+    '#34495E',
+    '#2E4053',
+    '#283747',
+    '#212F3C',
+    '#1B2631',
+    '#ABB2B9',
+    '#808B96',
+    '#566573',
+    '#2C3E50',
+    '#273746',
+    '#212F3D',
+    '#1C2833',
+    '#17202A'
+]
