+++ src/app/treasure-hunt/treasure-hunt-report/executive-summary/team-summary-pie-chart/team-summary-pie-chart.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, Input, ViewChild, ElementRef } from '@angular/core';
-import { graphColors } from '../../../../phast/phast-report/report-graphs/graphColors';
+import { graphColors } from '../../../../shared/graphColors';
 import * as _ from 'lodash';
 import { PlotlyService } from 'angular-plotly.js';
 @Component({
