+++ src/app/treasure-hunt/treasure-hunt-report/opportunity-payback/opportunity-payback-bar-chart/opportunity-payback-bar-chart.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit, Input, ElementRef, ViewChild } from "@angular/core";
 import { OpportunitiesPaybackDetails } from "../../../../shared/models/treasure-hunt";
-import { graphColors } from "../../../../phast/phast-report/report-graphs/graphColors";
+import { graphColors } from "../../../../shared/graphColors";
 import { Settings } from '../../../../shared/models/settings';
 import { PlotlyService } from "angular-plotly.js";
